+++ app/[locale]/projects/[projectId]/dmp/[dmpid]/s/[sid]/q/[qid]/__tests__/page.spec.tsx
@@ -7,6 +7,7 @@ import {
   useAnswerByVersionedQuestionIdQuery,
   usePublishedQuestionQuery,
   usePlanQuery,
+  useMeQuery
 } from '@/generated/graphql';
 import {
   addAnswerAction,
@@ -18,7 +19,7 @@ import * as apolloClientModule from '@/lib/graphql/client/apollo-client';
 import mockAnswerData from '../__mocks__/mockAnswerData.json';
 import mockPlanData from '../__mocks__/mockPlanData.json';
 import mockPublishedQuestion from '../__mocks__/mockPublishedQuestion.json';
-
+import mockMeData from '../__mocks__/mockMeQuery.json';
 // Mocked question data
 import mockCheckboxQuestion from '../__mocks__/mockCheckboxQuestion.json';
 import mockQuestionDataForBoolean from '@/__mocks__/common/mockPublishedQuestionDataForBoolean.json';
@@ -59,19 +60,29 @@ import { mockScrollIntoView } from "@/__mocks__/common";
 import PlanOverviewQuestionPage from "../page";
 import { AffiliationSearchQuestionType } from "@dmptool/types";
 
+// Mock for useComments hook
+import { mockUseComments, defaultMockReturn } from '../hooks/__mocks__/useComments';
+
+jest.mock('../hooks/useComments', () => {
+  const { mockUseComments } = jest.requireActual('../hooks/__mocks__/useComments');
+  return {
+    useComments: mockUseComments,
+  };
+});
+
 beforeEach(() => {
   // Cannot get the escaping to work in the mock JSON file, so doing it programmatically here
   const affiliationQuery = 'query Affiliations($name: String!){ ' +
     'affiliations(name: $name) { ' +
-      'totalCount ' +
-      'nextCursor ' +
-      'items { ' +
-        'id ' +
-        'displayName ' +
-        'uri ' +
-      '} ' +
+    'totalCount ' +
+    'nextCursor ' +
+    'items { ' +
+    'id ' +
+    'displayName ' +
+    'uri ' +
     '} ' +
-  '}';
+    '} ' +
+    '}';
 
   const json: AffiliationSearchQuestionType = {
     type: 'affiliationSearch',
@@ -99,6 +110,9 @@ beforeEach(() => {
     },
   };
   mockQuestionDataForTypeAheadSearch.publishedQuestion.json = JSON.stringify(json);
+
+  // Mock the useComments hook
+  mockUseComments.mockReturnValue(defaultMockReturn);
 });
 
 expect.extend(toHaveNoViolations);
@@ -106,6 +120,7 @@ expect.extend(toHaveNoViolations);
 
 // Mock the GraphQL query and mutation hooks
 jest.mock("@/generated/graphql", () => ({
+  useMeQuery: jest.fn(),
   usePlanQuery: jest.fn(),
   usePublishedQuestionQuery: jest.fn(),
   useAnswerByVersionedQuestionIdQuery: jest.fn(),
@@ -157,6 +172,14 @@ describe('PlanOverviewQuestionPage render of questions', () => {
 
     jest.useFakeTimers();
 
+    (useMeQuery as jest.Mock).mockReturnValue({
+      data: mockMeData,
+      loading: false,
+      error: undefined
+    });
+
+    mockUseComments.mockReturnValue(defaultMockReturn);
+
     (usePlanQuery as jest.Mock).mockReturnValue([
       jest.fn().mockResolvedValueOnce(mockPlanData),
       { loading: false, error: undefined },
@@ -173,6 +196,7 @@ describe('PlanOverviewQuestionPage render of questions', () => {
       loading: false,
       error: undefined,
     });
+
   })
 
   afterEach(() => {
@@ -199,7 +223,8 @@ describe('PlanOverviewQuestionPage render of questions', () => {
       );
     });
 
-    // Check for Requirements content
+
+    //Check for Requirements content
     expect(screen.getByRole('heading', { level: 3, name: 'page.requirementsBy' })).toBeInTheDocument();
     const boldedRequirements = screen.getByText('Requirements - Lorem Ipsum');
     expect(boldedRequirements).toBeInTheDocument();
@@ -222,7 +247,7 @@ describe('PlanOverviewQuestionPage render of questions', () => {
     expect(screen.getByRole('button', { name: 'Required by funder' })).toBeInTheDocument();
     // check for drawer panel trigger buttons
     expect(screen.getByRole('button', { name: 'page.viewSampleAnswer' })).toBeInTheDocument();
-    expect(screen.getByRole('button', { name: '4 Comments' })).toBeInTheDocument();
+    expect(screen.getByRole('button', { name: 'buttons.commentWithNumber' })).toBeInTheDocument();
     // Check that the textArea question field is in page
     expect(screen.getByLabelText('question-text-editor')).toBeInTheDocument();
 
@@ -266,7 +291,7 @@ describe('PlanOverviewQuestionPage render of questions', () => {
     expect(screen.getByRole('heading', { level: 2, name: 'Checkbox question' })).toBeInTheDocument();
     // View sample text button should not display when the question is not a textArea question type
     expect(screen.queryByRole('button', { name: 'page.viewSampleAnswer' })).not.toBeInTheDocument();
-    expect(screen.getByRole('button', { name: '4 Comments' })).toBeInTheDocument();
+    expect(screen.getByRole('button', { name: 'buttons.commentWithNumber' })).toBeInTheDocument();
     const checkboxGroup = screen.getByTestId('checkbox-group');
     expect(checkboxGroup).toBeInTheDocument();
     const checkboxes = within(checkboxGroup).getAllByRole('checkbox');
@@ -311,7 +336,7 @@ describe('PlanOverviewQuestionPage render of questions', () => {
     expect(screen.getByRole('heading', { level: 2, name: 'Radio button question' })).toBeInTheDocument();
     // View sample text button should not display when the question is not a textArea question type
     expect(screen.queryByRole('button', { name: 'page.viewSampleAnswer' })).not.toBeInTheDocument();
-    expect(screen.getByRole('button', { name: '4 Comments' })).toBeInTheDocument();
+    expect(screen.getByRole('button', { name: 'buttons.commentWithNumber' })).toBeInTheDocument();
 
     //Radio group
     const radioGroup = screen.getByRole('radiogroup');
@@ -350,7 +375,7 @@ describe('PlanOverviewQuestionPage render of questions', () => {
     expect(screen.getByRole('heading', { level: 2, name: "Testing" })).toBeInTheDocument();
     // View sample text button should not display when the question is not a textArea question type
     expect(screen.queryByRole('button', { name: 'page.viewSampleAnswer' })).not.toBeInTheDocument();
-    expect(screen.getByRole('button', { name: '4 Comments' })).toBeInTheDocument();
+    expect(screen.getByRole('button', { name: 'buttons.commentWithNumber' })).toBeInTheDocument();
     const textFieldWrapper = screen.getByTestId('field-wrapper');
     expect(textFieldWrapper).toBeInTheDocument();
     expect(within(textFieldWrapper).getByLabelText('text')).toBeInTheDocument();
@@ -385,7 +410,7 @@ describe('PlanOverviewQuestionPage render of questions', () => {
     expect(screen.getByRole('heading', { level: 2, name: 'Affiliation search question' }))
     // View sample text button should not display when the question is not a textArea question type
     expect(screen.queryByRole('button', { name: 'page.viewSampleAnswer' })).not.toBeInTheDocument();
-    expect(screen.getByRole('button', { name: '4 Comments' })).toBeInTheDocument();
+    expect(screen.getByRole('button', { name: 'buttons.commentWithNumber' })).toBeInTheDocument();
     const typeAheadContainer = screen.getByTestId('typeaheadWithOther');
     expect(typeAheadContainer).toBeInTheDocument();
     expect(within(typeAheadContainer).getByText('Institution')).toBeInTheDocument();
@@ -416,7 +441,7 @@ describe('PlanOverviewQuestionPage render of questions', () => {
     expect(screen.getByRole('heading', { level: 2, name: 'Date range question' })).toBeInTheDocument();
     // View sample text button should not display when the question is not a textArea question type
     expect(screen.queryByRole('button', { name: 'page.viewSampleAnswer' })).not.toBeInTheDocument();
-    expect(screen.getByRole('button', { name: '4 Comments' })).toBeInTheDocument();
+    expect(screen.getByRole('button', { name: 'buttons.commentWithNumber' })).toBeInTheDocument();
     // Date range container
     const dateRangeContainer = screen.getByTestId('date-range-container');
     expect(dateRangeContainer).toBeInTheDocument();
@@ -467,7 +492,7 @@ describe('PlanOverviewQuestionPage render of questions', () => {
     expect(screen.getByRole('heading', { level: 2, name: 'Date field question' })).toBeInTheDocument();
     // View sample text button should not display when the question is not a textArea question type
     expect(screen.queryByRole('button', { name: 'page.viewSampleAnswer' })).not.toBeInTheDocument();
-    expect(screen.getByRole('button', { name: '4 Comments' })).toBeInTheDocument();
+    expect(screen.getByRole('button', { name: 'buttons.commentWithNumber' })).toBeInTheDocument();
     // Date picker container
     const datePicker = screen.getByTestId('date-picker');
     const date = within(datePicker);
@@ -505,7 +530,7 @@ describe('PlanOverviewQuestionPage render of questions', () => {
     expect(screen.getByRole('heading', { level: 2, name: 'Yes/No question' })).toBeInTheDocument();
     // View sample text button should not display when the question is not a textArea question type
     expect(screen.queryByRole('button', { name: 'page.viewSampleAnswer' })).not.toBeInTheDocument();
-    expect(screen.getByRole('button', { name: '4 Comments' })).toBeInTheDocument();
+    expect(screen.getByRole('button', { name: 'buttons.commentWithNumber' })).toBeInTheDocument();
 
     //Radio group
     const radioGroup = screen.getByRole('radiogroup');
@@ -539,7 +564,7 @@ describe('PlanOverviewQuestionPage render of questions', () => {
     expect(screen.getByRole('heading', { level: 2, name: 'Url question' })).toBeInTheDocument();
     // View sample text button should not display when the question is not a textArea question type
     expect(screen.queryByRole('button', { name: 'page.viewSampleAnswer' })).not.toBeInTheDocument();
-    expect(screen.getByRole('button', { name: '4 Comments' })).toBeInTheDocument();
+    expect(screen.getByRole('button', { name: 'buttons.commentWithNumber' })).toBeInTheDocument();
 
     const urlInput = screen.getByPlaceholderText('url');
     expect(urlInput).toBeInTheDocument();
@@ -568,7 +593,7 @@ describe('PlanOverviewQuestionPage render of questions', () => {
     expect(screen.getByRole('heading', { level: 2, name: 'Email field question' })).toBeInTheDocument();
     // View sample text button should not display when the question is not a textArea question type
     expect(screen.queryByRole('button', { name: 'page.viewSampleAnswer' })).not.toBeInTheDocument();
-    expect(screen.getByRole('button', { name: '4 Comments' })).toBeInTheDocument();
+    expect(screen.getByRole('button', { name: 'buttons.commentWithNumber' })).toBeInTheDocument();
 
     const emailInput = screen.getByPlaceholderText('email');
     expect(emailInput).toBeInTheDocument();
@@ -598,7 +623,7 @@ describe('PlanOverviewQuestionPage render of questions', () => {
     expect(screen.getByRole('heading', { level: 2, name: 'Currency Field question' })).toBeInTheDocument();
     // View sample text button should not display when the question is not a textArea question type
     expect(screen.queryByRole('button', { name: 'page.viewSampleAnswer' })).not.toBeInTheDocument();
-    expect(screen.getByRole('button', { name: '4 Comments' })).toBeInTheDocument();
+    expect(screen.getByRole('button', { name: 'buttons.commentWithNumber' })).toBeInTheDocument();
 
     // Should see increment and decrement buttons
     const decreaseButton = screen.getAllByLabelText('Decrease');
@@ -633,7 +658,7 @@ describe('PlanOverviewQuestionPage render of questions', () => {
     expect(screen.getByRole('heading', { level: 2, name: 'Number Range question' })).toBeInTheDocument();
     // View sample text button should not display when the question is not a textArea question type
     expect(screen.queryByRole('button', { name: 'page.viewSampleAnswer' })).not.toBeInTheDocument();
-    expect(screen.getByRole('button', { name: '4 Comments' })).toBeInTheDocument();
+    expect(screen.getByRole('button', { name: 'buttons.commentWithNumber' })).toBeInTheDocument();
 
     expect(screen.getByText('Starting')).toBeInTheDocument();
     expect(screen.getByText('Ending')).toBeInTheDocument();
@@ -672,7 +697,7 @@ describe('PlanOverviewQuestionPage render of questions', () => {
     expect(screen.getByRole('heading', { level: 2, name: 'Number Field question' })).toBeInTheDocument();
     // View sample text button should not display when the question is not a textArea question type
     expect(screen.queryByRole('button', { name: 'page.viewSampleAnswer' })).not.toBeInTheDocument();
-    expect(screen.getByRole('button', { name: '4 Comments' })).toBeInTheDocument();
+    expect(screen.getByRole('button', { name: 'buttons.commentWithNumber' })).toBeInTheDocument();
 
     expect(screen.getByText('number')).toBeInTheDocument();
     // Should see increment and decrement buttons
@@ -708,7 +733,7 @@ describe('PlanOverviewQuestionPage render of questions', () => {
     expect(screen.getByRole('heading', { level: 2, name: 'Multi select question' })).toBeInTheDocument();
     // View sample text button should not display when the question is not a textArea question type
     expect(screen.queryByRole('button', { name: 'page.viewSampleAnswer' })).not.toBeInTheDocument();
-    expect(screen.getByRole('button', { name: '4 Comments' })).toBeInTheDocument();
+    expect(screen.getByRole('button', { name: 'buttons.commentWithNumber' })).toBeInTheDocument();
 
     expect(screen.getByText('Select Items (Multiple)')).toBeInTheDocument();
     expect(screen.getByText('Apple')).toBeInTheDocument();
@@ -748,7 +773,7 @@ describe('PlanOverviewQuestionPage render of questions', () => {
     expect(screen.getByRole('heading', { level: 2, name: 'Select box question' })).toBeInTheDocument();
     // View sample text button should not display when the question is not a textArea question type
     expect(screen.queryByRole('button', { name: 'page.viewSampleAnswer' })).not.toBeInTheDocument();
-    expect(screen.getByRole('button', { name: '4 Comments' })).toBeInTheDocument();
+    expect(screen.getByRole('button', { name: 'buttons.commentWithNumber' })).toBeInTheDocument();
 
     const selectButton = screen.getByTestId('select-button');
     expect(within(selectButton).getByText('Oregon')).toBeInTheDocument();
@@ -864,6 +889,14 @@ describe('Call to updateAnswerAction', () => {
       loading: false,
       error: undefined,
     });
+
+    (useMeQuery as jest.Mock).mockReturnValue({
+      data: mockMeData,
+      loading: false,
+      error: undefined
+    });
+
+    mockUseComments.mockReturnValue(defaultMockReturn);
   })
 
   it('should call updateAnswerAction when data changes for typeaheadSearch field', async () => {
@@ -1752,7 +1785,7 @@ describe('Call to updateAnswerAction', () => {
           versionedQuestionId: 'versionedQuestionId already exists'
         },
         id: 27,
-        json: "{\"type\":\"textArea\",\"answer\":\"This is a test\"}",
+        json: "{\"type\":\"textArea123\",\"answer\":\"This is a test\"}",
         modified: "1751929006000",
         versionedQuestion: {
           versionedSectionId: 20
@@ -1894,6 +1927,15 @@ describe('DrawerPanel', () => {
       loading: false,
       error: undefined,
     });
+
+    (useMeQuery as jest.Mock).mockReturnValue({
+      data: mockMeData,
+      loading: false,
+      error: undefined
+    });
+
+    mockUseComments.mockReturnValue(defaultMockReturn);
+
   })
 
   it('should open Sample text DrawerPanel when user clicks on the \'View sample answer\' button', async () => {
@@ -2039,7 +2081,7 @@ describe('DrawerPanel', () => {
     const sidebarPanel = screen.queryByTestId('sidebar-panel');
 
     // check for drawer panel trigger buttons
-    const viewCommentsBtn = screen.getByRole('button', { name: '4 Comments' });
+    const viewCommentsBtn = screen.getByRole('button', { name: 'buttons.commentWithNumber' });
     expect(viewCommentsBtn).toBeInTheDocument();
     expect(sidebarPanel).toBeInTheDocument();
     expect(visibleDrawerPanel).toBeUndefined();
@@ -2102,7 +2144,7 @@ describe('DrawerPanel', () => {
     });
 
 
-    const viewCommentsBtn = screen.getByRole('button', { name: '4 Comments' });
+    const viewCommentsBtn = screen.getByRole('button', { name: 'buttons.commentWithNumber' });
     // Click on Comments button to reveal comments drawer panel
     await userEvent.click(viewCommentsBtn);
 
@@ -2114,93 +2156,9 @@ describe('DrawerPanel', () => {
     await userEvent.click(commentBtn);
 
     // Get new info on sidebar and drawer panel - drawer should be closed and sidebar panel should be displayed
-    const allDrawerPanels2 = screen.getAllByTestId('drawer-panel');
-    const visibleDrawerPanel2 = allDrawerPanels2.find(
-      panel => panel.getAttribute('aria-hidden') !== 'true'
-    )!; // Non-null assertion operator
+
     const sidebarPanel2 = screen.queryByTestId('sidebar-panel');
     expect(sidebarPanel2).toBeInTheDocument();
-    expect(visibleDrawerPanel2).toBeUndefined();
-  })
-
-  it('should close drawer panel when user clicks on the masked body while comments drawer panel is open', async () => {
-
-    (usePublishedQuestionQuery as jest.Mock).mockReturnValue({
-      data: mockQuestionDataForTextArea,
-      loading: false,
-      error: undefined,
-    });
-
-    (useAnswerByVersionedQuestionIdQuery as jest.Mock).mockReturnValue({
-      data: mockAnswerDataForTextArea,
-      loading: false,
-      error: undefined,
-    });
-
-    await act(async () => {
-      render(
-        <PlanOverviewQuestionPage />
-      );
-    });
-
-
-    const viewCommentsBtn = screen.getByRole('button', { name: '4 Comments' });
-    // Click on Comments button to reveal comments drawer panel
-    await userEvent.click(viewCommentsBtn);
-
-    const layoutWithPanel = screen.getByTestId('layout-with-panel');
-
-    // Click on the masked layoutWithPanel to close the drawer
-    await userEvent.click(layoutWithPanel);
-
-    // Drawer should be closed now and sidebar panel should be displayed
-    const allDrawerPanels = screen.getAllByTestId('drawer-panel');
-    const visibleDrawerPanel = allDrawerPanels.find(
-      panel => panel.getAttribute('aria-hidden') !== 'true'
-    )!; // Non-null assertion operator
-    const sidebarPanel = screen.queryByTestId('sidebar-panel');
-    expect(sidebarPanel).toBeInTheDocument();
-    expect(visibleDrawerPanel).toBeUndefined();
-  })
-
-  it('should close drawer panel when user clicks on the masked body when sample answer drawer panel is open', async () => {
-
-    (usePublishedQuestionQuery as jest.Mock).mockReturnValue({
-      data: mockQuestionDataForTextArea,
-      loading: false,
-      error: undefined,
-    });
-
-    (useAnswerByVersionedQuestionIdQuery as jest.Mock).mockReturnValue({
-      data: mockAnswerDataForTextArea,
-      loading: false,
-      error: undefined,
-    });
-
-    await act(async () => {
-      render(
-        <PlanOverviewQuestionPage />
-      );
-    });
-
-
-    const viewSampleTextBtn = screen.getByRole('button', { name: 'page.viewSampleAnswer' });
-    // Click on View Sample text button to reveal sample text drawer panel
-    await userEvent.click(viewSampleTextBtn);
-
-    const layoutWithPanel = screen.getByTestId('layout-with-panel');
-
-    // Click on the masked layoutWithPanel to close the drawer
-    await userEvent.click(layoutWithPanel);
-
-    // Drawer should be closed now and sidebar panel should be displayed
-    const allDrawerPanels = screen.getAllByTestId('drawer-panel');
-    const visibleDrawerPanel = allDrawerPanels.find(
-      panel => panel.getAttribute('aria-hidden') !== 'true'
-    )!; // Non-null assertion operator
-    const sidebarPanel = screen.queryByTestId('sidebar-panel');
-    expect(sidebarPanel).toBeInTheDocument();
-    expect(visibleDrawerPanel).toBeUndefined();
   })
 });