+++ app/[locale]/styleguide/page.tsx
@@ -741,6 +741,7 @@ function Page() {
                 <DmpIcon icon="check_circle" />
                 <DmpIcon icon="error_circle" />
                 <DmpIcon icon="close" />
+                <DmpIcon icon="right-panel_close" />
                 <DmpIcon icon="solid-left_arrow" />
                 <DmpIcon icon="solid-right_arrow" />
                 <DmpIcon icon="solid-down_arrow" />